+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 * Fix autohold not applying consistently
 * Fix removing multiple ram watches
 * After a savefile is renamed, flag the old savefile as removed
+* Support writes past the end of custom SaveFile streams
 
 ## [1.4.6] - 2024-07-05
 ### Added